test_requirements.txt
pyCrypto
nose
